screen_process.c: Move the USB processing functions out of ui_screen_process.
First step towards modularization of the firmware code.

The calls to usb_processing_process() are done within a separate
main loop, which alternatively spins ui_screen_process once and
processes USB. This is quite hack-ish for now as it implies an ad-hoc
"is_done" function for spinning the UI only once, but separating
responsibilities between the UI and USB layers is more important if
we want to reach really modular code.