ui/components/trinary_input: add a special chars keyboard
Since there are 33 special chars (including the space), this means
there are now up to 11 chars per group. As a result, the line break in
a group is now after 6 chars, not after 5, so all of them can fit on
two rows.

Also added a pixel of space between the two character rows at the
bottom (and moved the actual password line up by the same).