releases: add script to verify a signed releases
The current instructions tell you how to reproduce the unsigned
firmware.  This adds a script to dump info about the signed release,
so the user can verify that the unsigned firmware is part of the
signed firmware binary.

It also dumps the hash which should match the hash shown by the
bootloader before booting the firmware.