workflow/unlock: visual optional passphrase confirmation
Instead of entering it twice, the passphrase is shown visually. This
is based on user feedback that entering a long passphrase twice, after
entering the device password, takes too long.

We do not have the underbracket in normal fonts yet, so the visual
confirmation shows spaces as normal spaces. For this reason, warnings
are added if there is a space at the beginning or end, as they are
impssible to see. Consecutive spaces also invoke a warning because it
might have been an accident and is also easy to overlook.