trinary_input_char.c: Fix the behaviour of deletions.
After inserting the first letter into a trinary_input_char,
the following letters would always require backspace to be pressed
twice in order to be effective, even in the case in which one was
in the main screen (i.e. not selecting a sub-alphabet).
This was due to the code setting the in_progress variable to true
even when a letter had just been selected.