qtouch: Use max(sensors) instead of sum(sensors) for touch detection.
Using the sum of sensor readings to determine whether the slider is
being touched or not is bad, as it depends on the location of the
finger w.r.t. the slider. Using max(readings) gives a much flatter
output, so applying a single touch/no touch threshold to it is easier.