gestures.c: Allow simultaneous slide and tap detection.
If the finger slides slightly on the sensor while performing
a touch, a tap event will still be emitted. This allows
both slides and taps through. Taps are considered taps
if and only if the finger didn't move beyond a certain
distance from the starting point during the duration of the gesture.