Enable R/W access to pages 2000-2047 in the MPU.
The default MPU configuration left by the bootloader disables
pages 2000-2047. This is pointless as it means losing 48 pages
at the end of the NVM main address space to no real benefit.

Extend the APP_DATA_1 MPU region until the end of the flash address
space; the protected bootloader section will remain R/W blocked for
the firmware as it has higher priority in the MPU table.