Py: Move noise verification callback after send, before response
The noise verification callback might not return if the callback
function does a blocking call to a client (for example to render the on
a window). If that is the case, the client would display the pairing
code before it is displayed on the device. To evade this, place the
callback after making the pairing verification request to the device,
and read the response once the callback returns.