cmake: Deduce PRODUCT_ macros from PLATFORM_ and EDITION_.
Instead of #defininig all the possible PRODUCT_ macros, it
makes more sense to enforce the definition of the (limited)
PLATFORM_ and EDITION_ macros, and extract the relevant definitions
of PRODUCT_* from there. This makes it easier to enforce that the macros
are defined consistently, and allows further separation of code that is
dependent on the platform it will run on, but not on the edition (e.g.
the touch code depends on the PLATFORM_*, but it doesn't matter whether
it will run on a -multi or on a -btconly device).