firmware: BTC sign, supported locktime and sequence bytes for RBF
BTC sign now supports nlocktime values between 0 and 500000000 (meaning
a nlocktime that actually refers to blocks and not median time).

nlocktime is only enabled in Bitcoin transactions that have a nsequence
number below MAXINT (0xffffffff). For this purpose we now also allow
0xffffffff-1 as a valid sequence number.

Further, electrum (for whose support this commit was originally
built), uses replace-by-fee (RBF) transactions. RBF is signalled by any
nsequence number that is below 0xffffffff - 1. Electrum only really
makes use of 0xffffffff - 2, so we additionally allow 0xffffffff-2 as a
valid nsequence number.

Both the locktime and the choice of RBF are displayed to the user, such
that he can verify the values during transaction signing. RBF
verification is currently only enabled for Bitcoin, since Litecoin does
not support it. The rbf verification is not a security feature, since it
is only shown for sequence numbers below 0xffffffff.

py: Remove locktime assertion from bitbox02 lib