keystore: erase seed from memory before creating a new one
The process of creating a new wallet is:

create seed (status: seeded) -> unlock -> create backup (status:
initialized).

After unlock, before create backup, the device could be unplugged, in
which case the process starts again.

There is an issue though if the process is started again without
reconnecting: the previous seed is still in memory, but a new seed is
created, so unlock runs into the "Seed has suddenly changed" abort.

The abort is a sanity check, but in this case, nothing bad would have
happened otherwise.