ugui: don't draw out of screen horizontally
UG_PutChar does a lot of looping and pixel drawing, which is pretty
slow. A longtouch confirm for example is noticable slower if a long
string is rendered, like an xpub (most of it cut off being out of
screen).

With a small bound check the device feels snappier and the
scrolling/confirm speeds are much more consistent.

Vertical check could be added, but we have no vertical scrolling, so
there is no benefit.