Py: Add firmware version checks
The communication lib now checks that the firmware is newer than a
minimum firmware version and raises an exception if this is not the
case.

Additionally a function is added in the bitbox02 to check wether the
firmware is newer than what is required as a minimum, meaning that
the library should probably be updated.