hww: graduate op_info to the hww framing layer
OP_INFO's purpose is to query the firmware version if it is not
available via the usb descriptor (bitboxbase, bitboxbridge, etc.), so
querying it should work the same way independently of the firmware
version.

Since the HWW framing layer was introduced, the call to OP_INFO needs
to be framed, which breaks the purpose. The fix is to graduate it to
the top level (treated as a special case).