Keystore: Use keystore to get xpub with simple xpub version bytes
The ethereum app had to call code in the bitcoin app to get the xpub
version bytes. This breaks the app isolation in the code. The ethereum
xpub call just requires the simple xpub, without version tweaks, so this
can be handled in a separate function in keystore.c . Also change this
for the electrum encryption key function, which now also serializes the
xpub directly in keystore.c .