py/bitbox02: direct pairing confirmation
In the BitBoxApp, MEW, etc., the flow so far has been:

1) Wallet shows pairing code, confirmation button is disabled
2) User confirms on device, which enables the OK button on the wallet
3) User confirms on the host.

The Python callback did not allow for this flow, which resulted in
e.g. the user being able to click continue in Electrum with no effect,
until the user also confirmed on the device.

This makes the `show_pairing` callback be able to perform the above
flow, so the host can show the pairing code, wait for the user
confirmation on the device, before proceeding to get the user
confirmation on the host.