rust: map CMake targets to rust features
For each target, we create one feature in Rust, which in turns defines
the set of features that make up that target.

Bootloader is removed as a platform, at the platforms we defined
previously were about the hardware (bitbox02, bitboxbase), not the
target on the platform. It can be readded as an additional dependent
feature when needed (same as for semihosting).