rust: use nightly-2020-03-25
This is to be able to use async/await syntax, which was merged a few
days ago into nightly for no_std:
https://github.com/rust-lang/rust/pull/69033.

This feature will be in beta in April and stable in June, at which
point we can start to use those channels again.

rust-toolchain file is added as a sanity check or for devs/users not
using Docker.