rust/bitbox02: add zero-on-drop Password struct
It is to be used with the C functions dealing with the password, and
has the same length.

No conversion happening to a Rust &str as only the null terminated C
bytes are currently used, so it would be just converted back and
forth. This could be adjusted in time.