ui/components/confirm: remove `shorten_body`, always truncate label
Shorten body was used to show the first and last 16 chars when
verifying an Ethereum data field or Ethereum signed message (if too
long).

This is arbitrary, there is no reason to prefer the last 16 bytes, or
to truncate at all. After all, it is strictly better to scroll through
a long string if you wanted to, than not being able to verify it at
all.