apps/btc/sign: replace 'See the BitBoxApp' flicker between outputs
The device waits for the next api call from the client after
confirming an output. During this time, the 'See the BitBoxApp' is
shown as the default screen if there is nothing else on the screen
stack.

This can result in an ugly flicker if there are multiple outputs to
confirm.

This is an attempt to reduce the flicker by just rendering an empty
screen (black screen) instead the regular waiting screen in this case.