font-size: convert screen
Unclear if this has any real effect in the firmware, as most
components that print text select the font again locally.

The bootloader uses the same init, and the strings are affected by
this. For now we force it back to 9x9 in the bootloader, as the
'upgrading' label looks more appealing in the original size.