CMakeLists,rust: enable RUSTC_BOOTSTRAP=1 also for docs
in bitbox02-rust-c, we use

Enabling lang-items is unstable.

The workaround is to use RUSTC_BOOTSTRAP=1 to enable this (while still
being on a stable toolchain), which was applied to the unit tests and
firmware, but not the docs.

I assume the docs target was added after we temporarily went to the
nightly toolchain.