apps/btc/sign: allow multiple script configs in a tx
Before, one script config (address type) was allowed for all inputs
and changes.

This commit changes the API of the signing workflow to allow
specifying multiple script configurations that can be used in a tx,
instead of only one.

In each output and change, the `script_config_index` now references
which script config to use.

The current commit still only allows only one (see TODO), but
implements the change on the API level. The TODO will be handled in a
subsequent commit.