noise: pop screens on new handshake
Not doing so is a regression from the rewrite of noise to Rust. This
removes the 'See the BitBoxApp' screen and goes to the default waiting
screen.

See
https://github.com/digitalbitbox/bitbox02-firmware/commit/d90d5e14207fcb1f79a69492c7cee78f71b2ea11#diff-ae8fe1d4fe4e70004f0f0ee3300406e8L189

In the future, we should consider showing the 'See the BitBoxApp'
screen again once the app is closed.

The screen stack pop was moved to the beginning of the handshake as
opposed to the end. This avoids a short flash of the 'See the
BitBoxApp' screen (as there are multiple api calls to finish the
pairing process), and we know the app (or a client) connected when a
handshake request comes in.