protobuf/backup: reduce sizes
The list backup response size. serialized, was up to 33404 bytes if
all fields were maxed out. This is an accident, as we can't even
transmit more than ~7.6k protobuf packets.

In practice, the size was smaller and could not reach those numbers,
because the id field in reality is 64 bytes (hex format of a 32 byte
hash), not up to 256 bytes as specified.

In addition, we reduce the number of max backups in the list of 100 to
50. 50 is still far more than reasonable, but puts the maximum number
of bytes of the response at 7103 bytes.