apps/btc: do not verify rbf&locktime if locktime is 0
RBF verification is not a security feature (as described in the
comment), but merely a sanity check, meant for advanced wallets like
Electrum.

Locktime on the other hand might be security relevant, when the
referenced block is in the future.

Until now, we verified the locktime if it is enabled anyway, also 0,
as it technically is considered. Practically however, a locktime of 0
has no effect, as the current block is always > 0.

We want to be able to opt-in to RBF with transactions made in the
BitBoxApp. Currently if doing so, the user would be prompted to verify

"Transaction is RBF. Locktime on block: XYZ"

Which is gibberisch to an average user. Especially the latter, a user
can not really verify.

We do a workaround and skip verification of RBF and locktime if the
locktime is zero, with the effect that the BitBoxApp can make RBF
transcations without disturbing the user with a complicated screen on
the device. This should have no security downside, and UX is better.