backup: fix call to _encode_backup_data, wrong tag
_encode_backup_data() is called manually to calculcate the length of
the serialization of the `data` field. Just passing
`BackupData_fields` is wrong though, as that points to the first field
in the BackupData message, not the `data` field.

Practically the length did not change as serializing the tag for the
first field or the data field resulted in the same length, but it was
very confusing to understand what was going on.

This will help when upgrading nanopb, which changes the fields
representation.