src/backup: remove useless loop
`backup_create()` makes three backups. It calls backup_fill() once to
get the backup it writes three times.

backup_fill() also has a loop with 3 iterations, but I can't see the
reason for it. The loop counter `i` is not referenced, and the loop
seems to do the exact same thing three times.

Removing the loop passes the tests (unit tests and manual tests).