rust: move config files to src/rust
It was in the root before because .cargo/config would not be resolved:

https://github.com/rust-lang/cargo/issues/2930

Changing the working directory fixes it.

Putting those files to src/rust is good to keep it close to the rust
code.

This also allows to add another Rust project in e.g. /tools, with a
different config.