workflow/unlock: remove use of workflow_stack
`workflow_unlock_and_handle_error()` is only used in password_check(),
which is blocking. So an async status display here does nothing. It
might be a leftover from when this function was called when unlocking
in u2f, where async is very important; but u2f is already using rust
async unlock.