ui: fix regression in confirm() regarding label truncating
When calling the Rust version of confirm(), the title/body were
assumed to be smaller than MAX_LABEL_SIZE. This was not a correct
assumption, the confirm component can take arbitary strings, and if
longer than MAX_LABEL_SIZE, it will truncate it.

This regression resulted in a panic for very long confirm bodies, such
as when confirming an ETH transaction with a long data field.