rust/async_usb: do not borrow executor state during poll
Currently, USB_TASK_STATE, the main executor state, is borrowed for
the whole duration of `spin()`, which includes the polling/execution
of the current task.

This prevents the state to be borrowed again during the execution of a
task. In a normal async runtime, there are things like `spawn(task)`
etc. that need to borrow and modify the executor state though.

The normal way to achieve that is to pop the task from the executor
state (often a task queue), poll it, and put it back in if it is not
finished yet. We do the same, but we have an enum with at most one
task instead of a task queue, so popping just means removing the task
from the enum.

Motivation: in a future PR we plan to add a `next_request()` to the
executor, which can be called from the currently running task. This
future will send a response to the host and wait for the next request
from the host.

This will allow sequential handling of api workflows that involve
multiple api calls from the host, e.g. BTC singing. A big state
machine keeping track of what api call we expect next can be simplied
a lot using `.await` instead.