rust/bitbox02: add stubs for testing
The bitbox02 safely wraps C functions. In Rust tests, C is not linked,
and even if it was, properly mocking the functions would be difficult.

We do a workaround instead: by adding the 'testing' feature and
enabling it only in tests, we can supply stub/mock implementations
that can be used in unit tests, not referencing any C functions.

In future commits, we can add some light mocking infrastructure to be
able to check args and mock the return values.

Alternatives dismissed:
- Add a trait for memory, keystore, ui and mock those.
  This resulted in all the code littered with generic params and lots
  of boilerplate.
- Implement the C funtions in Rust using cdylib and use that:
  Too complicated compared to the testing feature.