components/trinary_input_string: autocomplete input to word
BIP39 words are identified by a prefix of 3 of 4 chars, e.g. `aba`
uniquely identifies `abandon`. Currently in the input, the user is
already shown a restricted alphabet that can only lead to valid words,
but after `aba`, they have to tap once for every char in `ndon`.

This patch checks if there is the word is already uniquely identified
by the prefix, and if so, completes the word, making word input much
more efficient.

The back button still works as before, deleting one char at a time
only. A future commit could improve this to remove the same number of
letters, e.g. remove `ndon` when hitting the back button on `abandon`.