components/scroll_through_all_variants: fix spacing between entries
This component is used for:
- Showing all words in the mnemonic
- Choosing the correct word among 5 when confirming the mnemonic words

It is also perfectly capable of being a menu component, where the
entries are menu items, and `select` selects one of the entries.

The spacing between the words was a bit of a heuristic, adding
SCREEN_WIDTH/2, assuming that all words are small enough to not
overlap with this. Space between words of different width end up being
different.

This commit fixes this by adding the same space between all
components (20px, chosen by manual testing). This also works when this
component is used as a menu with longer menu entries, so there is
never an overlap between the entries.