workflow/restore_from_mnemonic: allow user to edit previous word
If going back to the previous word, the word edited before is preset,
so the user knows what they had entered before. Otherwise, if there
was a mistake a number of words back, one would not now when to stop
going back.

If the previous word was a simple typo, it is also easier to start
with the full word already preset.