apps/btc/sign: show progress when streaming inputs and prevtxs
When there is a large number of inputs or previous transactions with a
lot of inputs/outputs, streaming all that data can take a
while. Currently the device only shows the BitBox02 logo until after
the data has been streamed and the first output is shown to the user.

A progress bar shows the user that progress is being made, which
relieves anxiety for when it takes longer than expected.

Possible future improvements:

- do not show the progress bar until X milliseconds have passed - no
need to show it if the whole process is very quick.

- also show a proress bar when streaming the inputs in pass2 -
creating the signatures.