api/CheckBackup: return ERR_USER_ABORT on user abort
More fitting error code. This technically constitutes a breaking
change and by semver rules would require a major version bump. We skip
this effort as bitbox02-api-go and the BitBoxApp will work without a
change. There is a very small chance that any user of the bitbox02
Python package in the wild is affected by this.