apps/btc/sign: move progress step update to after the work is done
We start with progress 0 in init. Input 1 should not set start with 0
again, but already progress the progress bar after the work is done.

Same for the prevtx inputs and outputs.

This also makes it possible to naturally show a full 100% after all
inputs. Previously it would stop one step short of it.