apps/eth: better message signing verification
Messages to be verified are shown as ascii or as hex if there are
non-printable chars.

Previously, newlines would count as non-printable, so a message like
"hello\nworld" would be shown as hex.

This commit improves it to show one line at a time.

The whole thing is done in Rust as it is much more comfortable to
split a string into lines.