btc: add xpub formatting option to BTCRegisterScriptConfig
Most, if not all wallets except for Electrum use `xpub...` and
`tpub...` only for formatting xpubs. When registering a multisig
setup, so far the bitbox02 showed the xpubs as Zpub,Ypub,Vpub,Upub
depending on the network and script type.

This commit adds an option to force xpub or tpub in a backwards
compatible manner.