apps/btc: add param to sort xpubs before hashing
The hash is used to identify a multisig registration.

In multisig, the xpub ordering does not really matter, but our hashes
would change with a different order. This was not a real issue in
practice, as Electrum stores and sends the xpubs in the same order
every time.

In PSBT (at least in the way HWI parses it), xpubs are unordered.

This adds a param to the hashing function to sort the xpubs before
hashing, so that their order does not matter.