apps/btc/multisig: confirm script type and keypath during registration
Currently we allow only:

m/48'/0'/0'/1' for p2wsh-p2sh
m/48'/0'/0'/2' for p2wsh

Since we only support so few keypaths and script types, we could get
away with skipping verification of them.

We also want to whitelist m/48'/0'/0' for both script types, which is
the keypath used by Nunchuk for any script type. Then it becomes more
important to verify the script type and keypath that is being
registered.