qtouch: fix device lockup due to a race condition
There were occasional non-deterministic device freezes where
processing would stop completely.

Using gdb in semihosting, we saw that it gets stuck in
`qtm_measure_node()`, called from `qtm_samd51_ptc_handler()`.

Making the interrupt call a critical section, making the section
atomic, seems to fix the issue.

It is not 100% clear what was going on. We excluded that the interrupt
is running while the interrupt is running with an atomic counter. It
seems that something else (the qtouch timer interrupt or
qtouch_process (called form the mainloop) is stepping on the ADC0
interrupt. The more USB requests are processed, the more likely a
freeze was, which could mean that the usb requests make the execution
time of the qtouch interrupt longer, increasing the chance of a race
condition.