external: upgrade cryptoauthlib to v3.2.5
The library handling the secure chip.

https://github.com/digitalbitbox/cryptoauthlib/tree/v3.2.5

See also:

https://github.com/MicrochipTech/cryptoauthlib/wiki/Upgrading-to-v3.2#api-changes-between-v31x-and-v32x

There is a new param `word_address` that according to the docs above
is supposed to be sent as the first byte. We do this already manually
by sending 0x03:
https://github.com/digitalbitbox/bitbox02-firmware/blob/7ebb8450be8f3f9f3232f1407bfcaec4d22257e4/src/securechip/securechip.c#L134. The
actual `word_address` param as set by cryptoauthlib is always 0xFF as
that is the default
[here](https://github.com/digitalbitbox/cryptoauthlib/blob/v3.2.5/lib/calib/calib_execution.c#L302)
and not 0x03 as set
[here](https://github.com/digitalbitbox/cryptoauthlib/blob/v3.2.5/lib/calib/calib_execution.c#L325)
because we define iface type to be ATCA_CUSTOM_IFACE
[here](https://github.com/digitalbitbox/bitbox02-firmware/blob/7ebb8450be8f3f9f3232f1407bfcaec4d22257e4/src/securechip/securechip.c#L186). 0xFF
is basically a dummy value and defers to the custom
implementation (ATCA_CUSTOM_IFACE).

The cryptoauthlib driver code (which we currently don't use, but is a
good reference) also newly sends the word address before reading:
https://github.com/MicrochipTech/cryptoauthlib/blob/v3.2.5/lib/hal/hal_i2c_start.c#L230-L240. I
did not port this to our securechip.c/_receive function in this commit
as we didn't do it before, it seems to work without, and I could not
figure out the purpose of sending the word address before reading.