external: update libwally-core
To 46da3615b8e0837964f443d6b113ad500ea89f97.

This is a bit ahead of the official latest 0.8.0 release, to include
this commit, which adds secp256k1-zpk as a submodule instead of as a
source-copy inside libwally-core:

https://github.com/ElementsProject/libwally-core/commit/1da463088e64189d04c21073eb8861be1b717ad1

This will make it a lot easier for us use a custom fork of secp256k1,
which we might want to add anti nonce covert channel protection.

The new release also gets rid of all custom commits to libwally-core
and secp256k1, which are now all merged and released upstream.

The submodule points to our mirror at:
https://github.com/digitalbitbox/libwally-core/tree/bitbox02-firmware-20201213