api/restore/from_mnemonic: move time confirmation up
If it is not the first thing asked of the user, the time will be
inaccurate by the amount of time the user spent entering the recovery
words.

Also, it is not nice to have to abort at that step after all the
effort has been done to enter the words.