apps/btc/sign: add optional support for the anti-klepto protocol
Optional so that old firmware continues to work (no breaking change).

The anti-klepto protocol is invoked if the host app sends a host nonce
commitment along with the BTCSignInputRequest in the second pass. If
that commitment is present, instead of returning the signature, there
is one more roundtrip with the host:

1. The signer returns the signer commitment.
2. The host responds with the host nonce.
3. The signer signs the input using the host nonce contribution and
returns the signature.

If the host nonce commitment is not provided by the host the signature
is returned directly like previously, which ensures backwards
compatibility with old client libraries.

If the host nonce commitment is provided but an old firmware is used
that does not implement this change, the protobuf field is ignored as
it is unknown, which ensures backwards compatibility.