rust-test: Create a library with c and rust code
CMake is an absolutely horrible language. However I managed to write a
custom target to merge the library that is outputted from cargo with all
the object files that is "our" code into a library called
"bitbox_merged". Cargo is then told via a build-script to link to that
library (and our external dependencies) for tests. It seems to work \o/.