keystore: display unlock error code
Users reported a panic on the device after unlock: "keystore unlock
failed".

This was a panic because it was never supposed to happen. The expected
error (incorrect password) was caught and handled with a special
message. Since it is happening in for users anyway sometimes, we don't
panic anymore and we pass through granular error codes and display
them to the user, so we can get a better idea of what is going wrong.

The next commit will also display the securechip error code in case
of a securechip failure.