bootloader: make screen messages more user friendly
Some consider "bootloader" text on the screen confusing, so the
bootloader now invites users to follow instructions on the
BitBoxApp screen, consistent with the regular mode once a firmware is
booted.

In the case where app flash is empty, it also displays a "let's get
started" message, as well as rendering chunk writes progress with
"INSTALLING" label. BitBox02 is shipped with empty app flash. It can
also be erased manually via the bootloader protocol.

The commit also introduces a separate bootloader changelog since we've
never had it before and bootloader's release cycle is different from
that of the firmware.