ui/oled: Add support for new display SSD1312
The firmware and bootloader decide which driver to use based on a flag
in memory, "screen_type". The right value has to be set during factory
setup.

The bootloader could have had the value hardcoded, but it also looks
at the memory and chooses the driver dynamically based on that to
decouple the bootloader version from the hardware specifics (in
theory, new devices could use either screen type).

The bootloader version is bumped as a new bootloader is needed to
handle the new screens.

Co-authored-by: Marko Bencun <marko@shiftcrypto.ch>