commander: remove RandomNumber API request
No client uses it. The BitBoxApp was the only known user and dropped
the feature in v4.26.0.

Technically this constitutes a major version bump (breaking change),
but we cheat and only do a minor version bump to v9.7.0. This is a
shortcut because we are nearly certain that no client uses this
feature and will not notice the breaking change. A major bump would be
a big hassle as downstream integrations would need to be changed to
accept the new major version.

For the Python library, we can bump the major version without a
problem.