ethereum: allow mainnet on testnet keypath and vice versa
Until now:

Mainnet coins were under the xpub: m/44'/60'/0'/0
Testnet coins were under the xpub: m/44'/1'/0'/0

With this commit, one can use either path with any Ethereum
network. If it is not matching the above, a warning is displayed.

The reason this is changed is that some wallets, like Metamask, use
the 60' path for all networks. Using testnet coins on Metamask hence
fails since the BitBox02 rejected the transactions based on the
keypath.

Mixing keypaths like this is safe even if the user ignores the
warning, as the chain id (unique per ETH network) is part of the
transaction sighash, so one cannot be tricked into sending mainnet
coins instead of testnet coins. The only risk is that mainnet coins
might not be found if sent to the testnet keypath, hence the warning.