sd: remove _USE_LFN sections
_USE_LFN does not exist - FF_USE_LFN does. This code is not active and
can be removed.

LFN stands for long filename format, an extension to FAT32 that allows
filenames longer than what "8.3 filenames" allow (8 chars + 3 char
extension).

From http://elm-chan.org/fsw/ff/doc/readdir.html:

A short name is automatically generated from a long name according to
the FAT32 spec.

```
When LFN is enabled, a member altname[] is defined in the file
information structure to store the short file name of the object. If
the long file name is not accessible due to a reason listed below,
short file name is stored to the fname[] and the altname[] has a null
string.
```

I assume that in an old version of FatFs, the long filename was stored
in `lfname` and the short name in `fname`, while now the long name is
stored in `fname` and the short name is stored in `altname`.